pyproject.toml
[tool.poetry]
name = "openai-quickstart-python"
version = "0.1.0"
description = ""
authors = ["pythoninthegrass <4097471+pythoninthegrass@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "openai_quickstart_python"}]

[tool.poetry.dependencies]
python = "^3.11"
autopep8 = "1.6.0"
certifi = "2021.10.8"
charset-normalizer = "2.0.7"
click = "8.0.3"
et-xmlfile = "1.1.0"
flask = "2.0.2"
idna = "3.3"
itsdangerous = "2.0.1"
jinja2 = "3.0.2"
markupsafe = "2.0.1"
numpy = "^1.24.1"
openai = "0.19.0"
openpyxl = "3.0.9"
pandas = "1.3.4"
pandas-stubs = "1.2.0.35"
pycodestyle = "2.8.0"
python-dateutil = "2.8.2"
python-dotenv = "0.19.2"
pytz = "2021.3"
requests = "2.26.0"
six = "1.16.0"
toml = "0.10.2"
tqdm = "4.62.3"
urllib3 = "1.26.7"
werkzeug = "2.0.2"
python-decouple = "^3.7"

[tool.poetry.group.dev.dependencies]
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
